+++ CHANGES.md
@@ -1,3 +1,9 @@
+# 0.0.27
+
+* Add `Introspect.find_path` to retrieve steps from nested activities without having to
+  traverse them.
+* Use `trailblazer-activity-dsl-linear-1.0.0`.
+
 # 0.0.26
 
 * Fixing release, and allow using beta versions.